pyproject.toml
[project]
name = "yuuri-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.5.2",
    "flask>=3.1.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
